View filter for UU encoded data inside a message. WARNING: work in progress.
The code currently:
- detects the beginning of an UUencoded file,
- looks for the end of the file (from the end of the message, which
  means that several uuencoded files in the same message are not
  supported)
- replaces the whole UUencoded data by just a text like:

    UUencoded file named "foo.pdf"

So the current net value is (only) that the message is just a little
bit quicker to display (less text for the other filters). I don't
think this viewer removes any feature as, in any case, one needs to
save the message to a file and manually use uudecode to get the
file(s).

TODO:
One would need to code a new subclass of MimePart that would accept
raw text so that the viewer can build such a part and have it
displayed as an attachment.

And of course, do the decoding...

Feel free to jump on this, as I'm not sure at all I'll have time to
code them.