Make it an option (default on) that mark (un)read and many other commands automatically go to the next message.
Going to next message is a problem when messages are automatically viewed...
Suppose you have a set of unread messages in a folder. You open this folder.
This goes to the first unread message. You'd like to let it unread, so you hit
Ctrl-Shift-U. But then you have to repeat for the next message, and so on
until the end of the set of new messages...