Fix crash when connection to folder was lost during search.
mail_search_full() can realize that the connection to the mailbox was lost and
reset m_MailStream to NULL from inside it, so check that m_MailStream is still
valid after its return.

This could potentially explain a lot of crashes as M searches for unread/
deleted messages when closing the folder too.