Fix showing of encrypted messages with LF line separators.
Embedded encrypted messages may use only LF and not CR LF as line separators
but c-client functions expect the latter so ensure that we pass data in
correct format to them.

This required changing MimePartVirtual ctor to take wxCharBuffer instead of
wxMemoryBuffer and so also a few changes in uudecoding code.