Don't call wx ReleaseMouse() function as the mouse is not captured at wx level.
The workaround with forcefully breaking the mouse capture when showing dialogs
from TextViewerWindow::ProcessMouseEvent() resulted in asserts with recent wx
versions as the mouse was not captured at wx level, it was only done by the
native control itself.

Fix this by breaking the capture using Win32 function and not its wx
equivalent.