Use different icons for different frames under Windows.
We used the same icon for the main frame and all the other frames except the
log window which was confusing when many windows were opened. Also, the main
frame normal icon was completely different from the icon when new mail was
present.

Now use different, although similar, icons for each kind of frames, including
the composer and message view frames and the same kind of icon for the new
mail indication.

The new icons were taken from Gnome icons collection and slightly modified.
They are 48*48 now and not 32*32.

Finally change the program icon under Windows as well.