Don't use wxYield() when we just need to update our appearance.
The use of wxYield() in BusyIndicator and MProgressInfo could result in
crashes due to reentrancy in folder opening/header tables creation code. This
happened, in particular, when the folder was already being opened on resume
from suspend and a key was pressed in the meanwhile.

Simply don't use wxYield(), just calling Refresh() and Update() should be
enough to update the progress frame appearance and seems to work just fine at
least under Windows.