Don't save MFilterFromProfile lazily.
This is useless because the filter will need to be saved if Set() is called
and, conversely, it's only ever saved if it's called and so delaying saving it
complicates the code and results in bugs elsewhere, e.g. EditFilter() has code
setting the config source to use for saving the filter which didn't work as
expected because the filter was only saved in its dtor which was too late.