Correct the regular expression used for email detection.
The old one didn't work correctly at all as it detected bare email addresses
in the middle of the text so, provided a reasonably long text was present in
the clipboard, it was almost guaranteed to succeed. Anchor it properly now and
also fix the bug which now made it fail to detect the email addresses with
personal part that was uncovered by this change.