No real changes, just fix a lot of gcc const correctness warnings.
Fix all the warnings about converting string literals to non-const char*.

Most of these warnings occurred in the code using external libraries, either
c-client or Python, so add specific macros for casting the constness away
hoping that we might be able to remove them if the future versions of these
libraries become more const-correct.

Also remove a call provoking a warning from c-client linkage file directly in
the makefile when generating it.