Don't try to delete empty "temporary files".
MessageView::LaunchProcess() can be called with empty file name if the command
doesn't use any temporary files (which happens all the time when launching a
command to open an URL in a browser for example). Don't add empty file names
to the list of files to cleanup later, this is not necessary anyhow and avoids
an error when we try to delete a file with empty name later.