Send messages from a separate background thread in composer.
This avoids blocking the UI while the message is being sent which can take
quite some time for big messages and/or slow connections. Unfortunately saving
the message to SentMail (or Outbox, but this is less critical as Outbox is
normally a local folder) is still done in the main thread as currently folders
can't be accessed from other threads.

Split SendMessage::SendOrQueue() in PrepareForSending(), SendNow() and
AfterSending() with SendNow() being callable from non-main thread.

Create a thread in wxComposeView when sending the message and leave the window
opened but disabled until it terminates.