Avoid errors when marking as answered message in closed folder.
Trying to mark as answered a message in a folder which was closed while the
reply to it was composed results in assert failures and errors so don't even
try to do it, just warn the user about it.

Ideal would be to reopen the folder and still mark the message as answered but
this is not trivial to do as we don't have any MailFolder::Reopen() currently.