Don't monitor child process which we don't need to monitor.
We only need to monitor those child processes to which we pass names of
(temporary) files as the only reason for monitoring them is to delete these
files when the child process exits. So don't monitor the processes to which no
files are passed (main example in practice is opening URLs under Unix).

Also clean up the temporary files as soon as the child process terminates, not
only on program exit as before.