Split MMutex in STMutex and MTMutex, add MutexLocker.
Instead of defining MMutex class differently depending on whether we build
with USE_THREADS on or off, split it into two classes: STMutex and MTMutex
corresponding to the two versions. This allows to gradually start using
threads in some parts of the program instead of having to enable their use
everywhere (which would almost certainly break everything).

Also add MutexLocker class which is meant to replace the old MLocker which is
still kept for now until all of its uses can be examined and replaced with
either STMutex or MTMutex as appropriate.