Use a real mutex in Rfc822OutputRedirector.
Rfc822OutputRedirector can now be used from multiple threads, notably from a
worker thread sending the message and from the main thread if it tries to
autosave the composer contents, which uses SendMessage::WriteToFolder() that
uses Rfc822OutputRedirector in turn. Using a pseudo-mutex resulted in errors
in this case so use a real MTMutex here now instead.

This is not ideal as the main thread can now block trying to acquire this
mutex but this will be corrected later.