Don't keep alive folders opened non-interactively.
Keeping alive folders that are opened in the background is at best unnecessary
and at worst actively harmful as we can't even start the keep alive timer when
opening folder from another thread as timers are only supported in the main
thread by wxWidgets.

Also rename the vaguely named DoUpdate() to UpdateKeepAliveTimer() to describe
what it does more precisely and don't make it unnecessarily virtual.