Mention the "permanently keep alive" flag in the options dialog.
Mahogany UI is very confusing as you can either specify the delay for keeping
a folder open after it's closed by user in the "Folders" page or specify that
it should be always opened in the folders properties dialog "Access" page.
Ideal would be to somehow merge these two options but for now at least mention
the existence of the other one in the folders page help message.