Suppress error messages when IMAP server-side copy fails.
It's not an error from a user point of view if the client side still succeeded
(the simplest case when this happens is with servers that don't support
copying of messages into the same mailbox, such as Dovecot 1.2), so don't show
the error dialog in this case.

Do show the error as a status message and also show it as part of the error
message if the client-side copy fails as well.