Show hidden folders when their status changes.
If the folder status changes from inside Mahogany, it must be because a new
mail has been filtered to the folder and in this case it makes sense to show
it to the user so that the new mail could be seen even if the folder had been
previously hidden.

This allows to hide currently unused folders instead of removing them while
still keeping the filters for them as the folders would be shown again if
necessary (previously the filters had to be disabled to avoid potentially
losing new email).