Fix memory leak report for c-client home directory string under Windows.
We need to free myHomeDir c-client variable and not only myUserName one
manually as even if we hadn't used c-client at all, it can still be allocated.
OTOH we shouldn't free it in MailFolderCCCleanup() as this would result in a
double free. Finally we also don't need to set it in MailFolderCC as it
already seems to be set correctly by c-client anyhow.