Fix memory corruption bug when new mail arrived while opening a folder.
When a new mail arrived into an IMAP folder while we were opening it, memory
outside of the array we allocated could be accessed due to the mismatch
between the array size and the real number of messages in the folder.

This happened because the new message could arrive after we started sorting
the messages but before we threaded them -- in this case the threading code
used old, too small now, arrays and overwrote them with potentially fatal
consequences.

This long standing bug could also account for the problems with "phantom"
entries in the listing when a new mail arrived and was filtered away but this
is not certain as this bug is difficult to reproduce.