Wait for network connection reestablishment before reopening folders on resume.
Reopening remote folders immediately after resume fails under Windows 7
(unlike under XP) as the program receives the resume event before the network
connection is reestablished.

Fix this by waiting for a few (5 by default) seconds for the network to become
available before trying to reconnect. This is hackish and so the option to do
it is not even exposed in the UI for now but it does make resume work as
expected again.