Reimplement reopening folders after resume from sleep.
Instead of reopening the folders at the UI level, e.g. in the main frame
itself, reopen them at MailFolder level. This seems to make sense, can
potentially make things that are broken now (such as suspending with an option
reply in the composer and then sending the message after resuming won't mark
the original as answered) work in the future and fixes the bug when we were
losing connection to the folder reopened in the main frame on idle timeout
because we didn't restart its keep alive timer after reopening it.

Also make MailFolder::Close() public to allow calling it when suspending for
the folders that we don't need to keep open.