Ensure that the MailFolder opened in the main frame is associated with it.
Call MailFolder::SetInteractiveFrame() explicitly when opening a folder in the
main frame. While the interactive frame was set correctly when the folder was
opened the first time, it could remain NULL if a folder was opened, then the
view switched to another one and returned to the original folder without
reopening it because it remained open in the background because of the "keep
open" option set for it.