Enhance the "Quick filter" dialog by allowing to set up new folder from it.
Create the folder specified in the dialog rule if it doesn't exist and also
allow to set the new folder up sensibly by reusing the filter rule sender and
recipient.

This makes it as simple as possible to set up a new folder for all messages
from some sender which is probably the most common task for which the filters
are used.