Don't call MFPool::Remove() from MailFolderCmn dtor.
This is dangerous as if the folder was indeed left in the pool, due to some
bug, we're going to crash when calling MailFolder::GetName() from inside
MFPool::Remove() as it's pure virtual in MailFolderCmn and the derived object,
which did define it, is already destroyed by now.

So replace this check with a less precise but still good enough in practice
check for m_headers == NULL as if this is the case, the folder must have been
closed.