Fix filter parsing code in presence of non-ASCII characters.
We were getting asserts and wrong results when trying to read non-ASCII filter
rules. Fix this by using std::string instead of wxString for storing the
filter text as the code works with it as a sequence of bytes and not really as
characters. This is, of course, not ideal, especially because we currently
hard code the use of UTF-8 for converting wxString contents of the file into
std::string (which is fine in practice as long as the file is really encoded
in UTF-8) but better than nothing.