Fix crash when processing malformed inline-PGP-signed messages.
Verify that we don't go beyond the end of the message while searching for the
PGP signature. We actually didn't dereference any pointers to the memory
beyond the end of the message but we tried to construct a string with begin
position being greater than the end one in this case which used a negative
(i.e. huge positive when converted to unsigned) length and crashed.