Don't delete messages if reclassifying as spam fails.
Check for errors when reclassifying as spam and don't delete the messages that
couldn't be reclassified. At the very least this allows the user to have
another look at them in order to understand why reclassifying failed. And it
could also prevent data loss in case a wrong message was selected accidentally.