Simplify ClickableInfo mouse handling, drop double click action.
Double click action was always the same as single click one anyhow, except
that, confusingly, sometimes one was implemented in terms of another one and
sometimes vice versa. Simply remove OnDoubleClick() method and corresponding
WXMENU_LAYOUT_DBLCLICK constant.

Also, the mouse position was never used by the [double] click handler except
as the position to show the "Face" dialog at which was not very useful. So
remove the position argument from OnLeftClick() as it can be difficult to
obtain (e.g. for the upcoming wxWebView-based viewer).

Finally, as OnLeftClick() and OnRightClick() don't have the same signature any
more, don't use MessageView::DoMouseCommand() as a common interface for them
but just call them directly. This makes the code simpler and DoMouseCommand()
can be completely removed now.