Fix crash on exit with wxMSW 2.9.4 under Windows XP.
We deleted the tree window shown in the main frame but didn't remove it from
the splitter, so it was left with a dangling pointer. In wx 2.9.4, the
splitter receives an EVT_SIZE during the frame destruction and tries to use
these pointers resulting in a crash.

Fix this by calling Unsplit() before deleting the window.