Don't show folder in the main frame if we failed to reopen it.
The UI got into a confusing state if opening the folder shown in the main
frame after resume failed: it still showed the old folder contents and didn't
allow to reopen it (it had to be closed and reopened instead).

Fix this by clearing the main frame folder view in this case.