Remember attachment directories on a per folder basis.
Instead of having one global option, remember the directory used for selecting
the file to attach on a per-folder basis, as this seems more useful: different
directories would be typically used when writing to different recipients.

Also don't remember the name of the attachment at all, this doesn't seem to be
useful in practice as the same file is unlikely to be attached twice.