Don't set status of folders with empty MP_FTREE_FORMAT on startup.
We ignore changes to the status of these folders when the number of messages
in them changes later, but if we still set their status on startup, this can
result in inconsistent folder status, e.g. if a "Trash" folder (which is a
typical example of something you don't want to count messages in and so set
MP_FTREE_FORMAT to empty string for it) initially has any unread messages and
then is opened and cleared, it would still keep its "unread" colour.

To avoid this, while still not having to recount messages in the folder
whenever their number changes, don't set the status on startup neither. This
is not perfect, especially as it's still possible to get into inconsistent
state if the folder option is changed, but at least this state won't survive
after the program restart now.