Update Windows build instructions to reflect the current reality.
Update MSVC version used to build.

Replace mentions of svn with git.

Don't document building with Cygwin, this is not certain to work and is
probably not interesting to many people in practice anyhow. We definitely
shouldn't recommend using gcc 2.95 any more, this is just a wrong century for
it.