Slightly improve wxFont construction in CreateFontFromDesc().
There is no need to check for -1 font size as wxFontInfo already handles this
correctly by default, but OTOH there is no need to set the family if it is
default.

No real changes, this commit is just due to merging some of the old local
changes with the latest master.