Get rid of annoying variable shadowing warnings with MSVS 2015
Rename MFrameBase::name to m_name to avoid warnings whenever a very common
"name" variable was used anywhere in the code.

Make the scope of local variables narrower in several places and reuse the
existing variables in others.

Finally, use unique names if the variables are really different.

There should be no changes in behaviour except for one real (if minor) bug fix
in the update code where the message string was not filled correctly before.