Fix bug in display of nested parts in encrypted messages
The offset of the subparts was not computed correctly, it is relative to the
parent contents start and not the start of the body of this part itself.

This prevented Base64-encoded nested parts from being displayed at all in some
cases and didn't display the start of the message even for not encoded
messages.