Mostly ignore NOTATION_{NAME,DATA} in GPG output
This gets rid of the annoying warnings about unrecognized GPG output with the
notation data embedded into the signature.

It seems like this data can be mostly ignored, and only critical data, as
indicated by "NOTATION_FLAGS 1", should be shown, so attempt to do this,
although this is untested as I have no examples of signatures with critical
notation data.