Add a way to show log messages less intrusively
Create wxMLog class redirecting log messages to the currently active
wxInfoBar, if any, and set up an active wxInfoBar when showing a message in
a wxMessageView.

This shows any errors that may occur while parsing and showing the message in
this infobar without annoying the user with message boxes, which is much
better from the UI point of view.