Get rid of extra junk in quote prefix for mailing list posts
Many mailing lists use "From:" header of the form "Real Poster via
Mailing List" nowadays to work around DKIM constraints. This results in
using "RPvML" as quote prefix, which doesn't make much sense, so check
for this particular case and ignore everything coming after "via".

This is just a hack, but then so is the "via Mailing List" suffix to
begin with.