Allow viewing not quite UTF-8 messages using UTF-8
Ignore invalid characters when decoding the message text using UTF-8,
this can be useful when the message is mostly UTF-8, but just happens to
contain some trailing junk not using UTF-8 at the end.

And it's definitely better to show at least something when the user
selects "UTF-8" from the "Language" menu rather than nothing at all, as
happened before.