Prefer using lax UTF-8 decoding to Latin-1 for UTF-8 MIME parts
In 2019 it's vastly more likely that a message part using UTF-8 encoding
is actually encoded in UTF-8 and just has some non-UTF-8 bit appended to
it at the end, than it doesn't use UTF-8 at all (as happened 20 years
with some completely broken email-sending scripts that simple added
charset="UTF-8" to Content-Type without worrying about the actual
encoding of the data).

So stop replacing UTF-8 with Latin-1 at MimePart level and try to decode
the contents as UTF-8, but just ignore the non-UTF-8 characters instead.

From the user point of view, it means that the message with a non-UTF-8
trailer will now appear almost correctly by default, instead of having
to manually select "Language|UTF-8" to be able to see it correctly.