Decode PGP-encoded messages using UTF-8 instead of Latin-1
The decryption code seems to assume that anything coming back from gpg
is UTF-8 encoded which is probably not correct (to be tested with
encrypted messages using Content-Type charset different from UTF-8), but
as long as it does this, we need to use utf8_str() to work with decoded
data and not To8BitData().

This fixes showing encrypted messages with non-ASCII characters, which
previously couldn't be shown at all.