Fix wrong encoding conversion code in composer code
Passing "text" to cMB2WC() didn't make sense because this converted the
already Unicode string to multibyte using the current locale encoding,
which could fail and, in any case, seems completely useless.

Just ask the string itself for its multibyte representation in the
desired encoding instead.

This code is still overly complicated and the branches converting to the
part encoding and UTF-8 should be merged, but at least it works now.