Fix constructing "References" header in replies
This was broken since a very long time (and possibly ever) because
multiline "References:" in the original message were collapsed into a
single line, without any spaces between them.

Use HeaderIterator::MultiLineOk if the headers being retrieved includes
"References" to preserve the original header structure.

This is not ideal but at least doesn't mangle the header in the replies
any longer.