Fix comparing "From" addresses containing commas with own ones
Using AddressList was wrong here as it's supposed to be used with the
raw header values and the value we get here is already decoded, so
parsing it using the RFC 822 rules doesn't work correctly if it contains
any commas, for example, as they are address separators when decoded,
but can be included in the encoded form.

This fixes spurious warning messages about "Must use commas to separate
addresses" and wrong comparison results for the (decoded) address of the
form "Foo, Bar <foo@bar.com>".

This commit is best viewed ignoring whitespace-only changes.