Rewrite headers QP MIME encoding code once again
The previous version was still wrong as it didn't handle spaces between
words correctly any longer.

Simplify things by always encoding, instead of deciding whether to
encode for each word separately, this is not significantly less readable
in the encoded form (if we even still care about this) but much simpler
and more obviously correct.

Also perform our own QP-encoding instead of using c-client function,
it's so simple that it's much easier to do it like this.

Update the tests to match the results of the encoding algorithm now.