Strip literal CR LF from the internal header representation
Having CR LF in the "References:" header broke its value in the outgoing
messages after recent changes to MIME::EncodeHeader() which resulted in
not just throwing away all whitespace characters in the header when
encoding it.

It's not completely clear if AddHeaderEntry() should be given strings
with CR LF or if they should be stripped at some higher level, but for
now do it here to at least ensure that the program generates correct
outgoing messages.