Make code initializing SendMessageCC headers safer
Don't track the actual headers vector size in a separate variable, this
was error-prone and did result in bugs if the value of any header was
invalid because the final size wasn't set correctly then.

Instead, just keep the size correct at all times by appending the
elements to the vector instead of pre-resizing it first and then
resizing it again at the end.