Update Python wrappers code to SWIG 4.2 and Python 3
Rename SWIG-generated Mswigpyrun.h to Mswigpyrun.h-swig for consistency
with all the other SWIG-generated files.

Also remove the custom typemaps for String and just %apply the standard
string typemaps to it.

And take care of the macros not handled by SWIG in swigcmn.i itself
instead of doing it in the makefile.